stylised and worked date and time blocks
