+++ index.html
@@ -811,7 +811,7 @@
 
                                         <input type="date" name="date" id="input_date" value="" required>
 
-                                        <svg width="17" height="17" viewBox="0 0 17 17" fill="none" xmlns="http://www.w3.org/2000/svg" class="arrow_bottom">
+                                        <svg width="17" height="17" viewBox="0 0 17 17" fill="none" xmlns="http://www.w3.org/2000/svg" class="arrow_bottom" id="arrowBottomDate">
                                             <path d="M8.49028 2.00538e-06C6.80951 0.00192619 5.16704 0.502112 3.77048 1.43732C2.37393 2.37252 1.28597 3.70076 0.644144 5.25416C0.00231959 6.80755 -0.164562 8.51635 0.164589 10.1646C0.49374 11.8128 1.30415 13.3265 2.49338 14.5142C3.68261 15.7019 5.19727 16.5104 6.8459 16.8375C8.49454 17.1646 10.2031 16.9955 11.7557 16.3518C13.3083 15.708 14.6352 14.6184 15.5686 13.2206C16.5021 11.8229 17.0002 10.1798 17 8.49904C16.9994 7.38221 16.7787 6.27642 16.3507 5.24486C15.9227 4.2133 15.2957 3.27616 14.5055 2.48694C13.7153 1.69772 12.7773 1.0719 11.7452 0.64519C10.7131 0.218477 9.6071 -0.000764967 8.49028 2.00538e-06ZM8.80653 11.7161L4.11814 6.78347H13.4959L8.80653 11.7161Z" fill="#030303"/>
                                         </svg>
 
@@ -825,7 +825,7 @@
 
                                         <input type="time" name="time" id="input_time" required>
 
-                                        <svg width="17" height="17" viewBox="0 0 17 17" fill="none" xmlns="http://www.w3.org/2000/svg" class="arrow_bottom">
+                                        <svg width="17" height="17" viewBox="0 0 17 17" fill="none" xmlns="http://www.w3.org/2000/svg" class="arrow_bottom" id="arrowBottomTime">
                                             <path d="M8.49027 2.00538e-06C6.80951 0.00192619 5.16704 0.502111 3.77048 1.43732C2.37392 2.37252 1.28597 3.70076 0.644144 5.25415C0.00231959 6.80755 -0.164562 8.51634 0.164589 10.1646C0.493741 11.8128 1.30415 13.3265 2.49338 14.5142C3.68261 15.7019 5.19728 16.5104 6.84591 16.8375C8.49455 17.1646 10.2031 16.9955 11.7557 16.3518C13.3083 15.708 14.6352 14.6183 15.5686 13.2206C16.5021 11.8229 17.0002 10.1798 17 8.49904C16.9994 7.38221 16.7787 6.27642 16.3507 5.24486C15.9227 4.2133 15.2957 3.27616 14.5055 2.48694C13.7153 1.69772 12.7773 1.0719 11.7452 0.645189C10.7131 0.218476 9.6071 -0.000764967 8.49027 2.00538e-06ZM8.80653 11.7161L4.1201 6.78347H13.493L8.80653 11.7161Z" fill="#030303"/>
                                         </svg>
 
@@ -876,16 +876,11 @@
                                         </svg>
 
 
-                                        <select name="booking_ticket_type roboto" class="type_selector">
-                                            <option value="" disabled hidden>Ticket Type</option>
+                                        <select name="booking_ticket_type roboto" class="type_selector" tabindex="2">
                                             <option value="Permanent exhibition">Permanent exhibition</option>
                                             <option value="Temporary exhibition">Temporary exhibition</option>
                                             <option value="Combined Admission">Combined Admission</option>
                                         </select>
-
-                                        <svg width="17" height="17" viewBox="0 0 17 17" fill="none" xmlns="http://www.w3.org/2000/svg" class="arrow_bottom_ticket_type">
-                                            <path d="M8.49027 2.00537e-06C6.80951 0.00192619 5.16704 0.50211 3.77048 1.43731C2.37392 2.37251 1.28597 3.70075 0.644144 5.25414C0.00231959 6.80754 -0.164562 8.51633 0.164589 10.1646C0.493741 11.8128 1.30415 13.3264 2.49338 14.5142C3.68261 15.7019 5.19728 16.5104 6.84591 16.8375C8.49455 17.1646 10.2031 16.9955 11.7557 16.3517C13.3083 15.7079 14.6352 14.6183 15.5686 13.2206C16.5021 11.8229 17.0002 10.1798 17 8.49902C16.9994 7.38219 16.7787 6.27641 16.3507 5.24485C15.9227 4.21329 15.2957 3.27615 14.5055 2.48694C13.7153 1.69772 12.7773 1.0719 11.7452 0.645188C10.7131 0.218476 9.6071 -0.000764966 8.49027 2.00537e-06ZM8.80653 11.7161L4.1201 6.78345H13.493L8.80653 11.7161Z" fill="#030303"/>
-                                        </svg>
 
                                     </label>
                                 </div>
