+++ styles/1920px/section_tickets.css
@@ -350,7 +350,7 @@
     background: no-repeat;
     background-image: url(/assets/img/parallax.jpg);
     background-size: 100%;
-    background-position: 0 -150px;
+    background-position: 0 -144px;
     background-attachment: fixed;
     
 }
